+++ calculator/src/main/java/com/iteriam/calculator/service/OperatorService.java
@@ -5,5 +5,5 @@ import java.math.BigDecimal;
 public interface OperatorService {
 
 	
-	double calculateResult(BigDecimal parameter1,BigDecimal parameter2, String operator);
+	BigDecimal calculateResult(BigDecimal parameter1,BigDecimal parameter2, String operator);
 }
